Reduce unit-test execution time
