+++ integration/src/main/java/ai/djl/integration/tests/model_zoo/ModelZooTest.java
@@ -12,34 +12,19 @@
  */
 package ai.djl.integration.tests.model_zoo;
 
-import ai.djl.Application;
-import ai.djl.MalformedModelException;
 import ai.djl.Model;
 import ai.djl.ModelException;
-import ai.djl.basicmodelzoo.BasicModelZoo;
-import ai.djl.basicmodelzoo.cv.classification.AlexNet;
-import ai.djl.integration.util.TestUtils;
-import ai.djl.modality.Classifications;
-import ai.djl.modality.cv.Image;
-import ai.djl.modality.cv.output.DetectedObjects;
 import ai.djl.ndarray.NDList;
-import ai.djl.ndarray.types.Shape;
-import ai.djl.nn.Block;
 import ai.djl.repository.Artifact;
 import ai.djl.repository.zoo.Criteria;
 import ai.djl.repository.zoo.ModelLoader;
-import ai.djl.repository.zoo.ModelNotFoundException;
 import ai.djl.repository.zoo.ModelZoo;
 import ai.djl.repository.zoo.ZooProvider;
-import ai.djl.training.DefaultTrainingConfig;
-import ai.djl.training.Trainer;
-import ai.djl.training.loss.SoftmaxCrossEntropyLoss;
 import ai.djl.util.Utils;
 import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.ServiceLoader;
-import org.testng.Assert;
 import org.testng.SkipException;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -63,7 +48,7 @@ public class ModelZooTest {
     @Test
     public void testDownloadModels() throws IOException, ModelException {
         if (!Boolean.getBoolean("nightly") || Boolean.getBoolean("offline")) {
-            return;
+            throw new SkipException("Nightly only");
         }
 
         ServiceLoader<ZooProvider> providers = ServiceLoader.load(ZooProvider.class);
@@ -83,56 +68,6 @@ public class ModelZooTest {
                     }
                     Utils.deleteQuietly(Paths.get("build/cache"));
                 }
-            }
-        }
-    }
-
-    @Test
-    public void testImperativeModelInputOutput()
-            throws MalformedModelException, ModelNotFoundException, IOException {
-        // Test imperative models, only available in MXNet engine
-        if (!TestUtils.isMxnet()) {
-            throw new SkipException("Resnet50-cifar10 model only available in MXNet");
-        }
-        // from model zoo
-        Criteria<Image, Classifications> criteria =
-                Criteria.builder()
-                        .optApplication(Application.CV.IMAGE_CLASSIFICATION)
-                        .setTypes(Image.class, Classifications.class)
-                        .optGroupId(BasicModelZoo.GROUP_ID)
-                        .optArtifactId("resnet")
-                        .optFilter("layers", "50")
-                        .optFilter("dataset", "cifar10")
-                        .build();
-        try (Model model = ModelZoo.loadModel(criteria)) {
-            Assert.assertEquals(model.describeInput().values().get(0), new Shape(1, 3, 32, 32));
-            Assert.assertEquals(model.describeOutput().values().get(0), new Shape(1, 10));
-        }
-
-        Criteria<Image, DetectedObjects> ssdCriteria =
-                Criteria.builder()
-                        .optApplication(Application.CV.OBJECT_DETECTION)
-                        .setTypes(Image.class, DetectedObjects.class)
-                        .optGroupId(BasicModelZoo.GROUP_ID)
-                        .build();
-        try (Model model = ModelZoo.loadModel(ssdCriteria)) {
-            Assert.assertEquals(model.describeInput().values().get(0), new Shape(32, 3, 256, 256));
-            Assert.assertEquals(model.describeOutput().values().get(0), new Shape(1, 5444, 4));
-            Assert.assertEquals(model.describeOutput().values().get(1), new Shape(32, 5444, 2));
-            Assert.assertEquals(model.describeOutput().values().get(2), new Shape(32, 21776));
-        }
-
-        // from builder
-        Block alexNet = AlexNet.builder().build();
-        try (Model model = Model.newInstance("alexnet")) {
-            model.setBlock(alexNet);
-            try (Trainer trainer =
-                    model.newTrainer(new DefaultTrainingConfig(new SoftmaxCrossEntropyLoss()))) {
-                Shape inputShape = new Shape(32, 3, 224, 224);
-                trainer.initialize(inputShape);
-                Assert.assertEquals(
-                        model.describeInput().values().get(0), new Shape(32, 3, 224, 224));
-                Assert.assertEquals(model.describeOutput().values().get(0), new Shape(32, 10));
             }
         }
     }
